Fix handling of spaces in upstream file paths (#48)
File paths can have spaces in them, and they should not be escaped as if
they were URLs. This changes it so the URL parsing is done on the path
from the URL, before it is resolved into a full file path.
If the directory you are working in has a space in the path anywhere
upstream, npm test will fail. This commit fixes that problem. For
example, if you are on Windows working in:
C:\Users\me\My Documents\mywebpage\
then postcss-url with {url: "copy"} will fail due to handling of the
space in "My Documents".
Also, if you run npm test on postcss-url, and there is a space upstream,
that will also fail. For example, if you have cloned into:
C:\Users\me\My Documents\GitHub\postcss-url